test(code_executors): run container executor tests in CI across languages
Move the real-Docker ContainerCodeExecutor tests from tests/e2e into
tests/integration and add cases for javascript, typescript and shell
alongside python. Add a dedicated ubuntu-latest workflow that builds a
multi-runtime image and runs them with ADK_RUN_DOCKER_IT=1, so they now
execute in CI instead of only when opted in locally.
